Move app updater
