--- src/lib/github-utils/github-asset.d.ts
@@ -1,4 +1,0 @@
-type GithubAsset = {
-  name: string
-  url: string
-}
